SERVER-65831 Use ScopedTaskExecutor to speed up BalancerCommandsScheduler stop()
(cherry picked from commit c10fb605ba2b9d21882f4a98cc4f4314332acb1e)

+++ src/mongo/db/s/balancer/balancer_commands_scheduler_impl.cpp
@@ -175,7 +175,8 @@ void BalancerCommandsSchedulerImpl::start(OperationContext* opCtx,
     stdx::lock_guard<Latch> lg(_mutex);
     invariant(!_workerThreadHandle.joinable());
     if (!_executor) {
-        _executor = Grid::get(opCtx)->getExecutorPool()->getFixedExecutor();
+        _executor = std::make_unique<executor::ScopedTaskExecutor>(
+            Grid::get(opCtx)->getExecutorPool()->getFixedExecutor());
     }
     auto requestsToRecover = rebuildRequestsFromRecoveryInfo(opCtx, defaultValues);
     _numRequestsToRecover = requestsToRecover.size();
@@ -420,7 +421,7 @@ CommandSubmissionResult BalancerCommandsSchedulerImpl::_submit(
     }
 
     auto swRemoteCommandHandle =
-        _executor->scheduleRemoteCommand(remoteCommand, onRemoteResponseReceived);
+        (*_executor)->scheduleRemoteCommand(remoteCommand, onRemoteResponseReceived);
     return CommandSubmissionResult(params.id, distLockTaken, swRemoteCommandHandle.getStatus());
 }
 
@@ -465,7 +466,8 @@ void BalancerCommandsSchedulerImpl::_applyCommandResponse(
 
 void BalancerCommandsSchedulerImpl::_performDeferredCleanup(
     OperationContext* opCtx,
-    const stdx::unordered_map<UUID, RequestData, UUID::Hash>& requestsHoldingResources) {
+    const stdx::unordered_map<UUID, RequestData, UUID::Hash>& requestsHoldingResources,
+    bool includePersistedData) {
     if (requestsHoldingResources.empty()) {
         return;
     }
@@ -475,7 +477,7 @@ void BalancerCommandsSchedulerImpl::_performDeferredCleanup(
         if (request.holdsDistributedLock()) {
             _distributedLocks.releaseFor(opCtx, request.getNamespace());
         }
-        if (request.requiresRecoveryCleanupOnCompletion()) {
+        if (includePersistedData && request.requiresRecoveryCleanupOnCompletion()) {
             deletePersistedRecoveryInfo(dbClient, request.getCommandInfo());
         }
     }
@@ -537,7 +539,8 @@ void BalancerCommandsSchedulerImpl::_workerThread() {
         // 2.a Free any resource acquired by already completed/aborted requests.
         {
             auto opCtxHolder = cc().makeOperationContext();
-            _performDeferredCleanup(opCtxHolder.get(), completedRequestsToCleanUp);
+            _performDeferredCleanup(
+                opCtxHolder.get(), completedRequestsToCleanUp, true /*includePersistedData*/);
             completedRequestsToCleanUp.clear();
         }
 
@@ -565,17 +568,21 @@ void BalancerCommandsSchedulerImpl::_workerThread() {
         }
     }
     // Wait for each outstanding command to complete, clean out its resources and leave.
-    stdx::unordered_map<UUID, RequestData, UUID::Hash> requestsToClean;
+    (*_executor)->shutdown();
+    (*_executor)->join();
+
+    stdx::unordered_map<UUID, RequestData, UUID::Hash> cancelledRequests;
     {
         stdx::unique_lock<Latch> ul(_mutex);
-        _stateUpdatedCV.wait(
-            ul, [this] { return (_requests.size() == _recentlyCompletedRequestIds.size()); });
-        requestsToClean.swap(_requests);
+        cancelledRequests.swap(_requests);
         _requests.clear();
         _recentlyCompletedRequestIds.clear();
+        _executor.reset();
     }
     auto opCtxHolder = cc().makeOperationContext();
-    _performDeferredCleanup(opCtxHolder.get(), requestsToClean);
+    // Ensure that the clean up won't delete any request recovery document (the commands will be
+    // reissued once the scheduler is restarted)
+    _performDeferredCleanup(opCtxHolder.get(), cancelledRequests, false /*includePersistedData*/);
 }
